+++ include/triton/Conversion/TritonToTritonGPU/Passes.h
@@ -1,15 +1,15 @@
 #ifndef TRITON_CONVERSION_PASSES_H
 #define TRITON_CONVERSION_PASSES_H
 
-#include "triton/Conversion/TritonToTritonGPU/TritonToTritonGPUPass.h"
+#include "mlir/Pass/Pass.h"
 
-namespace mlir {
-namespace triton {
+namespace mlir::triton {
 
+#define GEN_PASS_DECL
+#include "triton/Conversion/TritonToTritonGPU/Passes.h.inc"
 #define GEN_PASS_REGISTRATION
 #include "triton/Conversion/TritonToTritonGPU/Passes.h.inc"
 
-} // namespace triton
-} // namespace mlir
+} // namespace mlir::triton
 
 #endif
