+++ include/triton/Conversion/TritonToTritonGPU/Passes.td
@@ -14,7 +14,6 @@ def ConvertTritonToTritonGPU: Pass<"convert-triton-to-tritongpu", "mlir::ModuleO
       type with an appropriate layout encoding (these encodings generally
       include information on `numWarps`, `threadsPerWarp` and `numCTAs`).
     }];
-    let constructor = "mlir::triton::createConvertTritonToTritonGPUPass()";
 
     let dependentDialects = ["mlir::arith::ArithDialect",
                              "mlir::math::MathDialect",
@@ -24,23 +23,34 @@ def ConvertTritonToTritonGPU: Pass<"convert-triton-to-tritongpu", "mlir::ModuleO
                              "mlir::triton::gpu::TritonGPUDialect"];
 
    let options = [
-       Option<"numWarps", "num-warps",
-              "int32_t", /*default*/"4",
-              "number of warps">,
-
-       Option<"threadsPerWarp", "threads-per-warp",
-              "int32_t", /*default*/"32",
-              "number of threads per warp">,
-        Option<"numCTAs", "num-ctas",
-              "int32_t", /*default*/"1",
-              "number of ctas in a cga">,
-        Option<"target", "target",
-              "std::string", /*default*/"\"\"",
-              "the GPU target, e.g., cuda:80, hip:gfx942">,
-        Option<"enableSourceRemat", "enable-source-remat",
-               "bool", /*default*/"false",
-               "enable trivial source rematerialization">,
+      Option<"target", "target",
+            "std::string", /*default*/"\"\"",
+            "the GPU target, e.g., cuda:80, hip:gfx942">,
+      Option<"numWarps", "num-warps",
+             "int32_t", /*default*/"4",
+             "number of warps">,
+      Option<"threadsPerWarp", "threads-per-warp",
+             "int32_t", /*default*/"32",
+             "number of threads per warp">,
+      Option<"numCTAs", "num-ctas",
+             "int32_t", /*default*/"1",
+             "number of ctas in a cga">,
+      Option<"enableSourceRemat", "enable-source-remat",
+             "bool", /*default*/"false",
+             "enable trivial source rematerialization">,
    ];
+}
+
+def ConvertTritonGPUToTritonGPU : Pass<"convert-tritongpu-to-tritongpu", "mlir::ModuleOp"> {
+  let summary = "relayout pass for `ttg` and `ttng` operations";
+  let description = [{
+    The `convert-tritongpu-to-tritongpu` pass is used during relayout of TTGIR
+    during warp specialization. Warp specialization may change the number of
+    warps for a partition, which requires reassigning layouts to all the
+    operations in the partition. However, those operations may include TritonGPU
+    and TritonNvidiaGPU dialect operations with specific layout requirements,
+    so they have to be re-inferred during this pass.
+  }];
 }
 
 #endif
